# rrule for BlockDiagonal * Vector multiplication