This is a draft, needs some love, thunking, and benchmarking.

- [x] love
- [x] thunking
- [x] benchmarking
- [x] bump patch

Putting it out for initial review.

One point in particular I want to emphasise is that the 
```
julia> b = BlockDiagonal([rand(2, 3), 2*ones(3, 3)])
5×6 BlockDiagonal{Float64,Array{Float64,2}}:
 0.127925  0.763018  0.590756  0.0  0.0  0.0
 0.21821   0.792263  0.696002  0.0  0.0  0.0
 0.0       0.0       0.0       2.0  2.0  2.0
 0.0       0.0       0.0       2.0  2.0  2.0
 0.0       0.0       0.0       2.0  2.0  2.0

julia> m = Matrix(b);

julia> v = rand(6);

julia> ȳ = rand(5);

julia> y, pb = rrule(*, m, v); unthunk(pb(ȳ)[2])
5×6 Array{Float64,2}:
 0.0412828  0.0166629  0.053694  0.0317837  0.0805506  0.0148633
 0.217186   0.0876624  0.28248   0.167212   0.42377    0.0781948
 0.452638   0.182698   0.588718  0.348487   0.883182   0.162966
 0.410036   0.165502   0.533308  0.315688   0.800058   0.147628
 0.435955   0.175964   0.567019  0.335643   0.85063    0.15696

julia> y, pb = rrule(*, b, v); unthunk(pb(ȳ)[2])
5×6 BlockDiagonal{Float64,Array{Float64,2}}:
 0.0412828  0.0166629  0.053694  0.0       0.0       0.0
 0.217186   0.0876624  0.28248   0.0       0.0       0.0
 0.0        0.0        0.0       0.348487  0.883182  0.162966
 0.0        0.0        0.0       0.315688  0.800058  0.147628
 0.0        0.0        0.0       0.335643  0.85063   0.15696
```

derivative with respect to `BlockDiagonal` is not the same as the one when using a `Matrix` (the off-block-diagonal elements are zero). This removes the redundant calculation of derivatives w.r.t. the off-block-diagonal elements.