@willtebbutt 2021-02-09
This is looking good.

I wonder whether it would make more sense to treat a `BlockDiagonal` like any old `struct` type an make its cotangent `Composite`s? For example, the constructor [here](https://github.com/invenia/BlockDiagonals.jl/blob/503bb6336a4605ba8e766372994bb16d85f27b72/src/chainrules.jl#L3) explicitly expects a `Composite`, and it deal with the semantic weirdness with the off-diagonals of the cotangent being zero, when they wouldn't be if the matrix were dense.