@wytbella 2021-02-10
So, after this PR,  the off diagonal elements of the derivative with respect to `BlockDiagonal` will be zero? How is this bahaviour tested?

Also, if the derivative with respect to `BlockDiagonal` and densified BlockDiagonal is not the same. Does that mean our old calculation has some problem (not just a performance issue)?