@willtebbutt 2021-02-10
> So, after this PR, the off diagonal elements of the derivative with respect to BlockDiagonal will be zero?

Kind of. The best way to reason about this is to completely ignore the fact that `BlockDiagonal` is a matrix, and instead view it as a `struct` with some data that we run some algorithm on (hence my suggestion to implement this using a `Composite` rather than a `BlockDiagonal`). If you derive the rule in that context, there's no notion of the of-diagonal elements, so there's really no problem.

> How is this bahaviour tested?

See `test/chainrules.jl` -- the AD testing functionality is now really clean a lot of the time because Lyndon added the final touches to some functionality we've had in the works for a while.

> Also, if the derivative with respect to BlockDiagonal and densified BlockDiagonal is not the same. Does that mean our old calculation has some problem (not just a performance issue)?

No. This is AD weirdness going on -- again it's best reasoned about by not thinking of `BlockDiagonal` as a matrix, for the sake of AD. The programmes
```julia
D = BlockDiagonal(ds)
D * x
```
and
```julia
D = Matrix(BlockDiagonal(ds))
D * x
```
actually wind up giving the same result w.r.t. `ds` because the reverse-pass through `Matrix` ignores the off-diagonals anyway.