@wytbella 2021-02-10
> hence my suggestion to implement this using a Composite rather than a BlockDiagonal). If you derive the rule in that context, there's no notion of the of-diagonal elements, so there's really no problem.

Ah, that makes sense! thanks!

My origianl question about "test" was how do we test that the off-diagonal of the derivative is zero, as [test_rrule](https://github.com/JuliaDiff/ChainRulesTestUtils.jl/blob/340e21c3aa277f55609681337ad85702c8b6d302/src/testers.jl#L144) looks pretty generic. I guess it's not relevant anymore if we view BlockDiagonal as a struct...

> because the reverse-pass through Matrix ignores the off-diagonals anyway.

How was this achieved when `D` is a normal matrix?