@willtebbutt 2021-02-10
> How was this achieved when D is a normal matrix?

It's the `rrule` for `Matrix(::BlockDiagonal)`, which you can find [here](https://github.com/invenia/BlockDiagonals.jl/blob/503bb6336a4605ba8e766372994bb16d85f27b72/src/chainrules.jl#L8) :)