@mzgubic 2021-02-10
> My origianl question about "test" was how do we test that the off-diagonal of the derivative is zero, as test_rrule looks pretty generic. I guess it's not relevant anymore if we view BlockDiagonal as a struct...

`test_rrule` actually tests that the block diagonal elements are zero via finite differences. That said, one needs to define a `to_vec` method for `BlockDiagonal` in a way that puts only the on-diagonal elements in the vector - which was there already, I just moved the code