@mzgubic 2021-02-10
Done some very simple benchmarking:
```
using BenchmarkTools
using BlockDiagonals
using Zygote

const bsmall = BlockDiagonal([rand(2, 3), rand(3, 3)])
const msmall = Matrix(bsmall)
const vsmall = rand(6)

const blarge = BlockDiagonal([rand(2, 2) for i in 1:100])
const mlarge = Matrix(blarge)
const vlarge = rand(200)

for (l, B, M, V) in [(:small, bsmall, msmall, vsmall), (:large, blarge, mlarge, vlarge)]
    println("Running benchmark for $l Matrix")
    @btime Zygote.gradient((m, v) -> sum(m*v), $M, $V);
    println("Running benchmark for $l BlockDiagonal")
    @btime Zygote.gradient((m, v) -> sum(m*v), $B, $V);
end
```

Before this PR:
```
Running benchmark for small Matrix
  451.878 ns (7 allocations: 1.11 KiB)
Running benchmark for small BlockDiagonal
  809.674 ns (12 allocations: 1.47 KiB)
Running benchmark for large Matrix
  55.466 μs (9 allocations: 630.53 KiB)
Running benchmark for large BlockDiagonal
  42.548 μs (209 allocations: 339.23 KiB)
```

With this PR
```
Running benchmark for small Matrix
  421.362 ns (7 allocations: 1.11 KiB)
Running benchmark for small BlockDiagonal
  1.588 μs (25 allocations: 2.63 KiB)
Running benchmark for large Matrix
  53.536 μs (9 allocations: 630.53 KiB)
Running benchmark for large BlockDiagonal
  32.362 μs (417 allocations: 56.61 KiB)
```

It looks like it gives us 2x slowdown for small matrices and 25% improvement for large ones. It looks like allocations can be improved however.