@mzgubic 2021-02-10
With improved indexing
```
Running benchmark for small Matrix
  322.243 ns (7 allocations: 832 bytes)
Running benchmark for small BlockDiagonal
  1.185 μs (22 allocations: 2.13 KiB) <-----
Running benchmark for large Matrix
  55.211 μs (9 allocations: 630.53 KiB)
Running benchmark for large BlockDiagonal
  28.735 μs (414 allocations: 53.09 KiB) <-----
```
Dummy indexing (best that can be achieved by dropping indexing allocations)
```
Running benchmark for small Matrix
  317.158 ns (7 allocations: 832 bytes)
Running benchmark for small BlockDiagonal
  919.414 ns (16 allocations: 1.56 KiB) <-----
Running benchmark for large Matrix
  55.780 μs (9 allocations: 630.53 KiB)
Running benchmark for large BlockDiagonal
  26.506 μs (408 allocations: 47.84 KiB) <-----
```