@mzgubic 2021-02-10
Adding `InplaceableThunk` for the vector cotangent (https://github.com/invenia/BlockDiagonals.jl/pull/58/commits/fc957e49558f75202aace57259fbfc21eb5cd4a4)

Zygote example unchanged

but running:
```julia
julia> @btime ((y, pb) = rrule(*, bsmall, $vsmall); (_, dm, dv) = pb($ȳsmall); dm)
```
before:
```julia
  1.342 μs (24 allocations: 2.31 KiB)
```
after:
```julia
  970.700 ns (18 allocations: 1.73 KiB)
```