@mzgubic 2021-02-10
Adding `InplaceableThunk` for each of the blocks (https://github.com/invenia/BlockDiagonals.jl/pull/58/commits/29b0b0684bd37747d2686d260acfea37bb4a709c)

gives the following:
```julia
Running benchmark for small Matrix
  391.565 ns (7 allocations: 832 bytes)
Running benchmark for small BlockDiagonal
  1.030 μs (18 allocations: 1.92 KiB)
Running benchmark for large Matrix
  54.101 μs (9 allocations: 630.53 KiB)
Running benchmark for large BlockDiagonal
  20.907 μs (214 allocations: 42.97 KiB)
```
which is 50% more time for small matrices and 50% less time for large matrices. (compared to without this PR).